Fixing bugs

+++ database/migrations/2019_08_19_045425_create_posibility_recommendations_table.php
@@ -16,7 +16,7 @@ class CreatePosibilityRecommendationsTable extends Migration
         Schema::create('posibility_recommendations', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('description');
-            $table->unsignedInteger('posibility_id');
+            $table->unsignedBigInteger('posibility_id');
             $table->timestamps();
         });
